tickets-service/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.11.1